Add handleCors middleware
