+++ sonar-project.properties
@@ -14,4 +14,4 @@ sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 
 
 sonar.test.inclusions=src/**/*.test.ts
-sonar.coverage.exclusions=src/index.ts, src/**/*.d.ts, src/**/types.ts, src/server/startServer.ts, src/database/connectToDatabase.ts, src/setupTestDatabase.ts
+sonar.coverage.exclusions=src/index.ts, src/**/*.d.ts, src/**/types.ts, src/server/startServer.ts, src/database/connectToDatabase.ts, src/setupTestDatabase.ts, src/**/handleCors.ts
