+++ src/globals/statusCodes.ts
@@ -4,6 +4,7 @@ const statusCodes = {
   OK: 200,
   CREATED: 201,
   BAD_REQUEST: 400,
+  FORBIDDEN: 403,
   NOT_FOUND: 404,
   CONFLICT: 409,
   INTERNAL_SERVER_ERROR: 500,
